# Conflicts:
#   src/main/java/de/unipassau/sep19/hafenkran/clusterservice/model/ExperimentDetails.java

+++ src/main/java/de/unipassau/sep19/hafenkran/clusterservice/service/impl/ExecutionServiceImpl.java
@@ -3,6 +3,7 @@ package de.unipassau.sep19.hafenkran.clusterservice.service.impl;
 import de.unipassau.sep19.hafenkran.clusterservice.dto.ExecutionCreateDTO;
 import de.unipassau.sep19.hafenkran.clusterservice.dto.ExecutionDTO;
 import de.unipassau.sep19.hafenkran.clusterservice.dto.ExecutionDTOList;
+import de.unipassau.sep19.hafenkran.clusterservice.dto.ResultDTO;
 import de.unipassau.sep19.hafenkran.clusterservice.dto.StdinDTO;
 import de.unipassau.sep19.hafenkran.clusterservice.dto.UserDTO;
 import de.unipassau.sep19.hafenkran.clusterservice.exception.ResourceNotFoundException;
@@ -13,6 +14,7 @@ import de.unipassau.sep19.hafenkran.clusterservice.model.ExperimentDetails;
 import de.unipassau.sep19.hafenkran.clusterservice.repository.ExecutionRepository;
 import de.unipassau.sep19.hafenkran.clusterservice.repository.ExperimentRepository;
 import de.unipassau.sep19.hafenkran.clusterservice.service.ExecutionService;
+import de.unipassau.sep19.hafenkran.clusterservice.serviceclient.ReportingServiceClient;
 import de.unipassau.sep19.hafenkran.clusterservice.util.SecurityContextUtil;
 import io.kubernetes.client.ApiException;
 import lombok.NonNull;
@@ -26,7 +28,11 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
-import java.util.*;
+import java.util.Base64;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
 import java.util.regex.Pattern;
 
 @Slf4j
@@ -39,6 +45,8 @@ public class ExecutionServiceImpl implements ExecutionService {
     private final ExperimentRepository experimentRepository;
 
     private final KubernetesClient kubernetesClient;
+
+    private final ReportingServiceClient rsClient;
 
     @Value("${kubernetes.deployment.defaults.ram}")
     private long ramDefault;
@@ -54,9 +62,15 @@ public class ExecutionServiceImpl implements ExecutionService {
      */
     @Override
     public String retrieveLogsForExecutionId(@NonNull UUID id, int lines, Integer sinceSeconds, boolean withTimestamps) {
+        ExecutionDetails executionDetails = retrieveExecutionDetailsById(id);
+
+        if (!executionDetails.getStatus().equals(ExecutionDetails.Status.RUNNING)) {
+            return "Logs can only be retrieved for running executions!";
+        }
+
         final String logs;
         try {
-            logs = kubernetesClient.retrieveLogs(retrieveExecutionDetailsById(id), lines, sinceSeconds,
+            logs = kubernetesClient.retrieveLogs(executionDetails, lines, sinceSeconds,
                     withTimestamps);
         } catch (ApiException e) {
             throw new ResponseStatusException(HttpStatus.SERVICE_UNAVAILABLE, "There was an error while " +
@@ -85,8 +99,11 @@ public class ExecutionServiceImpl implements ExecutionService {
      */
     @Override
     public ExecutionDetails createExecution(@NonNull ExecutionDetails executionDetails) {
+        executionDetails.validatePermissions();
+
         final ExecutionDetails savedExecutionDetails =
                 executionRepository.save(executionDetails);
+
 
         log.info(String.format("Execution with id %s created",
                 savedExecutionDetails.getId()));
@@ -101,6 +118,8 @@ public class ExecutionServiceImpl implements ExecutionService {
     public ExecutionDTO terminateExecution(@NonNull UUID executionId) {
 
         ExecutionDetails executionDetails = getExecutionDetails(executionId);
+
+        executionDetails.validatePermissions();
 
         try {
             kubernetesClient.deletePod(executionDetails);
@@ -144,10 +163,9 @@ public class ExecutionServiceImpl implements ExecutionService {
     private ExecutionDetails retrieveExecutionDetailsById(@NonNull UUID id) {
         final Optional<ExecutionDetails> execution = executionRepository.findById(id);
         ExecutionDetails executionDetails = execution.orElseThrow(
-                () -> new ResourceNotFoundException(ExecutionDetails.class, "id",
-                        id.toString()));
+                () -> new ResourceNotFoundException(ExecutionDetails.class, "id", id.toString()));
         executionDetails.validatePermissions();
-        return executionDetails;
+        return execution.get();
     }
 
     /**
@@ -231,6 +249,8 @@ public class ExecutionServiceImpl implements ExecutionService {
                 executionRepository.findById(executionId).orElseThrow(
                         () -> new ResourceNotFoundException(ExecutionDetails.class, "id", executionId.toString()));
 
+        executionDetails.validatePermissions();
+
         if (status.equals(Status.FINISHED)) {
             executionDetails.setTerminatedAt(LocalDateTime.now());
         }
@@ -245,7 +265,9 @@ public class ExecutionServiceImpl implements ExecutionService {
     }
 
     private ExecutionDetails startExecution(@NonNull ExecutionDetails executionDetails) {
-        String podName = null;
+        executionDetails.validatePermissions();
+
+        final String podName;
         try {
             podName = kubernetesClient.createPod(executionDetails);
         } catch (ApiException e) {
@@ -346,9 +368,19 @@ public class ExecutionServiceImpl implements ExecutionService {
      */
     public ExecutionDetails getExecutionOfPod(@NonNull String podName, @NonNull UUID namespace) {
 
-        ExperimentDetails experiment = experimentRepository.findById(namespace).orElseThrow(() -> new ResourceNotFoundException(ExperimentDetails.class, "id",
-                namespace.toString()));
+        ExperimentDetails experiment = experimentRepository.findById(namespace).orElseThrow(
+                () -> new ResourceNotFoundException(ExperimentDetails.class, "id",namespace.toString()));
+
+        experiment.validatePermissions();
+
         return executionRepository.findByPodNameAndExperimentDetails(podName, experiment);
+    }
+
+    @Override
+    public void updatePersistedResults(@NonNull ExecutionDetails execution) {
+        byte[] results = getResults(execution.getId());
+        ResultDTO resultDTO = new ResultDTO(execution.getId(), execution.getOwnerId(), Base64.getEncoder().encodeToString(results));
+        rsClient.sendResultsToResultsService(resultDTO);
     }
 
     /**
@@ -358,6 +390,8 @@ public class ExecutionServiceImpl implements ExecutionService {
     public ExecutionDTO deleteExecution(@NonNull UUID executionId) {
 
         ExecutionDetails executionDetails = getExecutionDetails(executionId);
+
+        executionDetails.validatePermissions();
 
         if (executionDetails.getStatus().equals(Status.RUNNING)) {
             throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Can not delete executions in running");
